Level init script
