+++ Frontend League 2019/Assets/Scripts/LevelSetup.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class LevelSetup : MonoBehaviour
+{
+    public void SetupScene(int levelNumber) => Debug.Log($"Level {levelNumber} has been loaded");
+}
