+++ Frontend League 2019/Assets/Scripts/LevelSetup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60c129eff415e474c84d1df13e3253e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
