+++ Frontend League 2019/ProjectSettings/EditorBuildSettings.asset
@@ -4,5 +4,11 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes: []
+  m_Scenes:
+  - enabled: 1
+    path: Assets/Scenes/Menu.unity
+    guid: 05e0181a1eca44b1ca361531b35051fe
+  - enabled: 1
+    path: Assets/Scenes/Level.unity
+    guid: 28f946b6deb334aad8d04e8f93b57217
   m_configObjects: {}
